refactor: separate ProductBatch models and improve service layer responsibilities
- Extract ProductBatch models to dedicated file
- Move section update validation logic from handler to service
- Improve separation of concerns between layers
